Create UndertowServiceEteTest specific to dialogue endpoint result test
